pyproject.toml
[project]
name = "djangorest"
version = "0.1.0"
description = ""
authors = [
    {name = "Valdemar",email = "vovateslenko.1996.28@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
